Fix heavily inefficient prompt sampling
Been doing this in other projects but didn't implement it here for some reason. The more variants there are in a specific system prompt, the more entries it will take up in memory. If there's enough combinations, the amount of prompts will simply explode in value. Bad! Awful!

Instead, we select a "base" template (one of the entries from the lists provided into `PromptManager`) and iterate through the string, directly subbing the variant strings with one of the choices contained within them.

Some other changes:
- Tweak threshold for `SimilarityFilter`
- Add a few more prompt variations for generic GTI prompts
- Don't have a new `PromptManager` initialized at every step of the `NoRobotsInstructionFollowingTask`